requirements.txt
streamlit
openai
pillow
assemblyai